requirements.txt
requests
python-dotenv
rich
pip-audit